move the repl code into its own file
